@@ -12,7 +12,6 @@
     "url": "https://github.com/bashaus/jest-ado-reporter.git"
   },
   "license": "MIT",
-  "type": "commonjs",
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
   "files": [
@@ -27,21 +26,24 @@
   },
   "dependencies": {},
   "devDependencies": {
-    "@commitlint/cli": "17.6.5",
-    "@commitlint/config-conventional": "17.6.5",
+    "@commitlint/cli": "20.1.0",
+    "@commitlint/config-conventional": "20.0.0",
+    "@eslint/js": "^9.36.0",
     "@swc/core": "^1.13.5",
     "@swc/jest": "^0.2.39",
-    "@types/jest": "^29.5.2",
-    "@types/node": "^20.3.1",
-    "@typescript-eslint/eslint-plugin": "8.45.0",
-    "@typescript-eslint/parser": "8.45.0",
-    "eslint": "^8.42.0",
+    "@types/jest": "^30.0.0",
+    "@types/node": "^24.6.1",
+    "eslint": "^9.36.0",
+    "eslint-config-prettier": "^10.1.8",
     "eslint-plugin-jest": "^29.0.1",
     "eslint-plugin-prettier": "^5.5.4",
     "husky": "^9.1.7",
-    "jest": "^29.5.0",
+    "jest": "^30.2.0",
+    "jest-junit": "^16.0.0",
     "prettier": "3.6.2",
-    "sort-package-json": "^2.4.1",
-    "typescript": "^5.9.3"
+    "sort-package-json": "^3.4.0",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.9.3",
+    "typescript-eslint": "^8.45.0"
   }
 }
